+++ CHANGELOG.md
@@ -20,6 +20,7 @@ v1.3.0-dev
  * ons-alert-dialog: Added event handler attributes.
  * ons-carousel: Added event handler attributes.
  * ons-dialog: Added event handler attributes.
+ * ons-navigator: Added event handler attributes.
 
 v1.2.2
 ----